% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbr_species.R
\name{abbr_species}
\alias{abbr_species}
\title{Abbreviate species}
\usage{
abbr_species(data, species_var = "species")
}
\arguments{
\item{data}{Species dataset}

\item{species_var}{Current species variable name. Defalts to "species"}
}
\description{
Abbreviate species
}
