% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mgms.R
\name{add_mgms}
\alias{add_mgms}
\title{Add MGMS to a dataframe}
\usage{
add_mgms(data, site_var = "site", cpue_var = "cpue")
}
\arguments{
\item{data}{A dataframe}

\item{site_var}{The name of the column in \code{data} that contains the site information}

\item{cpue_var}{The name of the column in \code{data} that contains the Catch Per Unit Effort (CPUE) information}
}
\value{
A dataframe with an additional column containing the MGMS
}
\description{
This function adds a Multigear Mean Standardization (MGMS) column to a dataframe
}
\examples{
data <- data.frame(site = c("A", "A", "B", "B"),
                   cpue = c(5, 10, 15, 20))
add_mgms(data)
}
