% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Format table}
\usage{
format_table(
  data,
  species_var = "especie",
  measure_var = "medidas",
  sep = " - "
)
}
\arguments{
\item{data}{Species dataset}

\item{species_var}{The species variable name. Defaults to "especie"}

\item{measure_var}{The measures variable name. Defaults to "medidas"}

\item{sep}{Separator of measures. Defaults to " - "}
}
\description{
Format table
}
