% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table_cpue.R
\name{format_table_cpue}
\alias{format_table_cpue}
\title{Format table of CPUE into long format}
\usage{
format_table_cpue(data, first_sp_col = 3)
}
\arguments{
\item{data}{A data frame of CPUE with species as columns}

\item{first_sp_col}{The column number of the first species column}
}
\value{
A data frame in long format
}
\description{
This function takes a table of Catch Per Unit Effort (CPUE) with species as columns and formats it into long format
}
\examples{
# Create a sample data frame
df <- data.frame(
  site = c("A", "B", "C", "D"),
  gear = c(1, 1, 1, 1),
  sp1 = c(2, 3, 5, 6),
  sp2 = c(1, 4, 2, 7),
  sp3 = c(4, 5, 1, 8))

# Format the table of CPUE
format_table_cpue(df)
}
