% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUPcollection-package.R
\docType{package}
\name{NUPcollection-package}
\alias{NUPcollection}
\alias{NUPcollection-package}
\title{NUPcollection: R package for NUP Collection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Set of functions for internal use at NUP Collection.
}
\author{
\strong{Maintainer}: Bruno Stabile \email{bhmstabile@gmail.com} (\href{https://orcid.org/0000-0002-3524-537X}{ORCID})

Other contributors:
\itemize{
  \item Thiago Pedroso (\href{https://orcid.org/0000-0001-6678-3333}{ORCID}) [contributor]
}

}
\keyword{internal}
