% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_especies.R
\name{relatorio_especies}
\alias{relatorio_especies}
\title{Relatório de Espécies}
\usage{
relatorio_especies(dataset)
}
\arguments{
\item{dataset}{Um dataframe com as colunas "ordem", "familia", "genero" e "especie".}
}
\value{
Uma lista de espécies formatada para o relatório.
}
\description{
Cria uma lista de espécies pronta para o relatório, ordenada por ordem, família, gênero e espécie.
}
\examples{
dataset <- data.frame(
  ordem = c("Carnivora", "Carnivora", "Primates", "Primates"),
  familia = c("Felidae", "Canidae", "Hominidae", "Cebidae"),
  genero = c("Panthera", "Canis", "Homo", "Cebus"),
  especie = c("Panthera leo", "Canis familiaris", "Homo sapiens", "Cebus apella")
)
relatorio_especies(dataset)
}
