% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_example_data}
\alias{sp_example_data}
\title{Species example data}
\format{
A data frame with 6 rows and 6 columns:
}
\source{
Personal data
}
\usage{
sp_example_data
}
\description{
Species example data
}
\keyword{datasets}
