% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_compare.R
\name{species_compare}
\alias{species_compare}
\title{Species compare}
\usage{
species_compare(
  data,
  species_target,
  species_var = "especie",
  measure_var = "medidas"
)
}
\arguments{
\item{data}{Species dataset}

\item{species_target}{The target species to compare}

\item{species_var}{The species variable name. Defaults to "especie"}

\item{measure_var}{The measure variable name. Defaults to "medidas"}
}
\description{
Species compare
}
