% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_compare.R
\name{species_compare}
\alias{species_compare}
\title{Species compare}
\usage{
species_compare(data, species_target, measure_var = "medidas")
}
\arguments{
\item{data}{data}

\item{species_target}{species_target}

\item{measure_var}{measure_var}
}
\description{
Species compare
}
